pyproject.toml
[tool.poetry]
name = "democraci-geomapper"
version = "0.1.0"
description = "A Python project for mapping coordinates to jurisdictions using GeoPandas."
authors = ["Michal Babinski <mbabinski17@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
geopandas = "^0.10.2"
shapely = "^2.0.4"
requests = "^2.31.0"
rtree = "^1.2.0"
python-dotenv = "^1.0.1"
tabulate = "^0.9.0"
sqlalchemy = "^2.0.30"
psycopg2 = "^2.9.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
